Created a new method (function?) that creates and returns a part of an array.
NOTE: I need to put the function call somewhere where the program runs it to actually get it to do what it does. ALSO if you want it to display its values then print it out using a "cout" else it WILL return the value but nothing will happen in terms of the console/display.